edited navigation layout to storyvoard referncing
